chore: reorganize shmq module
